Cambio de algoritmo de trilateracion

+++ src/main/java/com/meli/quasar/domain/entities/Position.java
@@ -1,22 +1,22 @@
 package com.meli.quasar.domain.entities;
 
 public class Position {
-	private float X;
-	private float y;
-	public float getX() {
+	private double X;
+	private double y;
+	public double getX() {
 		return X;
 	}
-	public void setX(float x) {
+	public void setX(double x) {
 		X = x;
 	}
-	public float getY() {
+	public double getY() {
 		return y;
 	}
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	
-	public void setXY(float[] xy) {
+	public void setXY(double[] xy) {
 		if (xy.length > 0) {
 			this.X = xy[0];
 			this.y = xy[1];
